Remove outer directories
