+++ src/api/drello-api/board.ts
@@ -1,6 +1,21 @@
 import { drelloApiAxios } from "src/utils/axios/drello-api-axios";
 import { path } from "src/utils/url/drello-api";
-import { Board } from "src/types/outer/drello-api/board";
+
+export interface Board {
+  id: number;
+  title: string;
+  columns: Column[] | [];
+}
+
+export interface Column {
+  id: number;
+  title?: string;
+  cards?: Card[];
+}
+export interface Card {
+  id: number;
+  title: string;
+}
 
 export const getBoards = () =>
   drelloApiAxios.get<{ boards: Board[] }>(path.boards());
