+++ src/api/drello-api/sign-up.ts
@@ -1,6 +1,4 @@
 import { drelloApiAxios } from "src/utils/axios/drello-api-axios";
 import { path } from "src/utils/url/drello-api";
-import { SignUp } from "src/types/outer/drello-api/sign-up";
 
-export const signup = () =>
-  drelloApiAxios.get<{ signUp: SignUp }>(path.signUp());
+export const signup = () => drelloApiAxios.get<{ signUp: {} }>(path.signUp());
