+++ src/redux/domain/board.ts
@@ -1,7 +1,7 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import { getBoards } from "src/api/drello-api/board";
 import { Board as innerBoard } from "src/types/inner/board.g";
-import { Board as OuterBoard } from "src/types/outer/drello-api/board";
+import { Board as OuterBoard } from "src/api/drello-api/board";
 import { RootState } from "src/redux/root";
 
 interface BoardState {
