--- src/types/outer/drello-api/board.d.ts
@@ -1,7 +1,0 @@
-import { Column } from "src/types/outer/column";
-
-export interface Board {
-  id: number;
-  title: string;
-  columns: Column[] | [];
-}
