--- src/types/outer/drello-api/card.d.ts
@@ -1,4 +1,0 @@
-export interface Card {
-  id: number;
-  title: string;
-}
