--- src/types/outer/drello-api/column.d.ts
@@ -1,7 +1,0 @@
-import { Card } from "src/types/outer/card.d";
-
-export interface Column {
-  id: number;
-  title?: string;
-  cards?: Card[];
-}
