--- src/types/outer/drello-api/sign-up.d.ts
@@ -1,1 +1,0 @@
-export interface SignUp {}
